Release 7.1

+++ classes/Formatter/Post/LastModifiedAuthor.php
@@ -10,12 +10,11 @@ use AC\Type\Value;
 
 class LastModifiedAuthor implements Formatter
 {
-
     public function format(Value $value): Value
     {
-        $user_id = get_post_meta($value->get_id(), '_edit_last', true);
+        $user_id = (int) get_post_meta($value->get_id(), '_edit_last', true);
 
-        if ( ! get_userdata($user_id)) {
+        if (! get_userdata($user_id)) {
             throw new ValueNotFoundException();
         }
